refactor: rename InputContainerSignUp files to unique name
